like feature finished

+++ client/xwitter/src/lib/supabase/mutations.ts
@@ -1,5 +1,6 @@
 "use server";
 
+import { revalidatePath } from "next/cache";
 import { supabaseServer } from ".";
 import { randomUUID } from "crypto";
 
@@ -14,14 +15,23 @@ export const likeTweet = async ({
     .from("likes")
     .insert({ id: randomUUID(), tweet_id: tweetId, user_id: userId });
 
+  revalidatePath("/");
   console.log(data, error);
 };
 
-export const unlikeTweet = async ({ likeId }: { likeId: string }) => {
+export const unlikeTweet = async ({
+  tweetId,
+  userId,
+}: {
+  tweetId: string;
+  userId: string;
+}) => {
   const { data, error } = await supabaseServer
     .from("likes")
     .delete()
-    .eq("id", likeId);
+    .eq("tweet_id", tweetId)
+    .eq("user_id", userId);
 
+  revalidatePath("/");
   console.log(data, error);
 };
